+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/right-panel/treeview/icons/DragHandle.tsx
